+++ MellomOppdrag/eksempel-app/src/main/webappplatform/angular-eksempel-app/src/app/employee/employee-list/employee-list.component.ts
@@ -47,17 +47,23 @@ export class EmployeeListComponent implements OnInit {
   }
 
   deleteEmployee(employee : Employee) : void {
-    this.service.deleteEmployee(employee.id)
-          .subscribe(
-              (data : Employee) => {
-                console.log(data);
-                this.allEmployees();
-              },
-              error => {
-                //this.createError = true;
-                console.log(error);
-              }
-          );
+    if (employee && employee.id) {
+      if (confirm(`Really delete the product: ${employee.firstName} ${employee.lastName}?`)) {
+
+        this.service.deleteEmployee(employee.id)
+            .subscribe(
+                (data : Employee) => {
+                  console.log(data);
+                  this.allEmployees();
+                },
+                error => {
+                  //this.createError = true;
+                  console.log(error);
+                }
+            );
+
+      }
+    }
   }
 
   public selectEmployee(employee : Employee) : void  {
